Moved graph functions to new module
